🔧 Fix: TCP/UDP visibility + Int/Long type mismatch
- PacketParser: TCP/UDP constants now public (used by CaptureVpnService)
- CaptureViewModel: sumOf { it.length } cast to Long for totalBytes

+++ app/src/main/java/com/packetlens/ui/viewmodel/CaptureViewModel.kt
@@ -118,7 +118,7 @@ class CaptureViewModel @Inject constructor(
             httpCount = allPackets.count { it.protocol == Protocol.HTTP || it.protocol == Protocol.HTTPS },
             dnsCount = allPackets.count { it.protocol == Protocol.DNS },
             tlsCount = allPackets.count { it.protocol == Protocol.TLS },
-            totalBytes = allPackets.sumOf { it.length }
+            totalBytes = allPackets.sumOf { it.length.toLong() }
         )
     }
